Type error fix
